Add support for textarea with floating_labels_form
Textarea doesn't currently work with the floating_labels_form simple_form wrapper. When you enter text, it displays under the label (which is also the front color and size).

Broken and fixed versions showed here: https://imgur.com/a/TR0IOym

Tested on Chrome and Firefox. Unable to test on Edge/IE, but I think the change made to the fallback for those browsers will work.